carecloud/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
google-generativeai
werkzeug
python-dotenv
Pillow
pytesseract
